abstract movement
