+++ modules/bindings/src/movement_bindings.cpp
@@ -73,7 +73,7 @@ void export_movement()
     def("accelerate_linear", accelerate_linear, (arg("from_speed"), arg("to_speed"), arg("accel_per_sec") = get_config().getDouble("linear_accel"), arg("updates_per_sec") = get_config().getInt("accelerate_updates_per_sec")));
     def("accelerate_sinusoidal", accelerate_sinusoidal, (arg("from_speed"), arg("to_speed"), arg("accel_per_sec") = get_config().getDouble("sinusoidal_accel"), arg("updates_per_sec") = get_config().getInt("accelerate_updates_per_sec")));
     def("gyro_drive_straight", gyro_drive_straight_fn, (arg("from_speed"), arg("to_speed"), arg("stop_function"), arg("correction_proportion") = get_config().getDouble("gyro_correction_proportion"), arg("accel_per_sec") = get_config().getDouble("gyro_linear_accel"), arg("updates_per_sec") = get_config().getInt("gyro_updates_per_sec")), "drive straight for `seconds` seconds using the gyroscope");
-    def("gyro_turn_degrees", gyro_turn_degrees, (arg("from_speed"), arg("to_speed"), arg("degrees"), arg("accel_per_sec") = get_config().getDouble("gyro_linear_accel"), arg("updates_per_sec") = get_config().getInt("gyro_updates_per_sec")), "turn a certain number of degrees");
+    def("gyro_turn_degrees", gyro_turn_degrees, (arg("turn_speed"), arg("degrees"), arg("updates_per_sec") = get_config().getInt("gyro_updates_per_sec")), "turn a certain number of degrees");
     def("gyro_turn_degrees_v2", gyro_turn_degrees_v2, (arg("max_speed"), arg("degrees"), arg("min_speed") = get_config().getInt("gyro_turn_v2_min_speed"), arg("accel_per_sec") = get_config().getDouble("gyro_linear_accel"), arg("updates_per_sec") = get_config().getInt("gyro_updates_per_sec")));
     def("is_black", is_black, (arg("val")));
     def("is_white", is_white, (arg("val")));
