+++ modules/movement/src/imu_movement.cpp
@@ -2,64 +2,9 @@
 #include "controllers.hpp"
 #include "accumulator.hpp"
 #include "config.hpp"
-#include <iostream>
+#include "singletons.hpp"
+#include "maneuvers.hpp"
 #include <kipr/wombat.h>
-using std::cout;
-using std::endl;
-
-#define between(val, a, b) ((a >= val && val >= b) || (b >= val && val >= a))
-
-// --------------------------------------------- GYRO SINGLETON ---------------------------------------------
-// Functions in this section deal with the Gyroscope Singleton
-// --------------------------------------------------------------------------------------------------------
-GyroSingleton::GyroSingleton(int updates_per_sec) : Accumulator(get_gyro_angle, updates_per_sec){};
-double GyroSingleton::get_gyro_val()
-{
-    signed short val = GYRO_FUNCTION();
-    return between(val, get_config().getDouble("min_gyro_val"), get_config().getDouble("max_gyro_val")) ? 0 : (static_cast<double>(val) - get_config().getDouble("mean_gyro_val")) * get_config().getDouble("gyro_cw_multiplier");
-}
-double GyroSingleton::get_gyro_angle()
-{
-    return get_gyro_val() / get_config().getDouble("raw_to_360_degrees");
-}
-GyroSingleton *GyroSingleton::instance()
-{
-    if (_instance.get() == nullptr)
-    {
-        _instance = std::shared_ptr<GyroSingleton>(new GyroSingleton());
-    }
-    return _instance.get();
-}
-
-std::shared_ptr<GyroSingleton> GyroSingleton::_instance = nullptr;
-
-// --------------------------------------------- GYRO SUBSCRIBER ---------------------------------------------
-// Functions in this section deal with the gyroscope subscriber class
-// --------------------------------------------------------------------------------------------------------
-GyroSubscriber::GyroSubscriber(int updates_per_sec) : start_angle(GyroSingleton::instance()->get_accumulator()),
-                                                      was_running(GyroSingleton::instance()->is_running())
-{
-    // setup gyro
-    GyroSingleton::instance()->set_updates_per_sec(updates_per_sec);
-
-    // start accumulating
-    GyroSingleton::instance()->start();
-}
-GyroSubscriber::~GyroSubscriber()
-{
-    if (!was_running)
-    {
-        GyroSingleton::instance()->stop();
-    }
-}
-const double &GyroSubscriber::get_start_angle()
-{
-    return start_angle;
-}
-double GyroSubscriber::get_relative_angle()
-{
-    return GyroSingleton::instance()->get_accumulator() - start_angle;
-}
 
 // --------------------------------------------- GYRO FUNCTIONS ---------------------------------------------
 // Functions in this section deal with gyroscope functions
@@ -100,79 +45,17 @@ void gyro_drive_straight(int from_speed, int to_speed, std::function<bool()> sto
     MOVEMENT_FUNCTION(to_speed, to_speed);
 }
 
-void gyro_turn_degrees(Speed from_speed, Speed to_speed, double degrees, double accel_per_sec, int updates_per_sec)
+void gyro_turn_degrees(Speed turn_speed, double degrees, int updates_per_sec)
 {
-    LinearController left_accelerator(from_speed.left, to_speed.left, accel_per_sec, updates_per_sec);
-    LinearController right_accelerator(from_speed.right, to_speed.right, accel_per_sec, updates_per_sec);
-
-    // setup gyro
     GyroSubscriber subscriber(updates_per_sec);
 
-    while ((degrees > 0 && subscriber.get_relative_angle() < degrees) || (degrees < 0 && subscriber.get_relative_angle() > degrees))
-    {
-        MOVEMENT_FUNCTION(static_cast<int>(left_accelerator.speed()), static_cast<int>(right_accelerator.speed()));
-        left_accelerator.step();
-        right_accelerator.step();
-        msleep(left_accelerator.get_msleep_time());
-    }
-    MOVEMENT_FUNCTION(to_speed.left, to_speed.right);
+    turn_degrees(&subscriber, turn_speed, degrees, updates_per_sec);
 }
 
 void gyro_turn_degrees_v2(int max_speed, double degrees, int min_speed, double accel_per_sec, int updates_per_sec)
 {
-    // this function is meant to deal with 2 main turning cases:
-    // 1 - where you accelerate to speed, have time to keep turning, then need to start
-    // decelerating after you reach the final_angle - angle_turned_while_accelerating.
-    // 2 - where you try to accelerate to speed, already reach half the angle, then need to start decelerating
-
-    LinearController accelerator(0, max_speed, accel_per_sec, updates_per_sec);
-
-    double cached_accumulator = 0;
-    double speed = accelerator.speed();
-    int left_sign = (degrees > 0) ? 1 : -1;
-    int right_sign = (degrees > 0) ? -1 : 1;
-
     // setup gyro
     GyroSubscriber subscriber(updates_per_sec);
 
-    // accelerating part of the turn, capped at 1/2 of the turn that way, if it didn't have enough time to accelerate,
-    // it will still have enough time to decelerate from its current speed
-    while ((degrees > 0 && subscriber.get_relative_angle() < degrees / 2) ||
-           (degrees < 0 && subscriber.get_relative_angle() > degrees / 2))
-    {
-        speed = accelerator.speed();
-        MOVEMENT_FUNCTION(static_cast<int>(speed * left_sign), static_cast<int>(speed * right_sign));
-        accelerator.step();
-        msleep(accelerator.get_msleep_time());
-
-        if (accelerator.done())
-        {
-            cached_accumulator = subscriber.get_relative_angle();
-            break;
-        }
-    }
-
-    // do any extra turning that is needed
-    MOVEMENT_FUNCTION(static_cast<int>(speed * left_sign), static_cast<int>(speed * right_sign));
-    while (cached_accumulator != 0 &&
-           ((degrees > 0 && subscriber.get_relative_angle() < degrees - cached_accumulator) ||
-            (degrees < 0 && subscriber.get_relative_angle() > degrees - cached_accumulator)))
-    {
-        msleep(accelerator.get_msleep_time());
-    }
-
-    // decelerate from current speed to min_speed, which should be close to 0
-    LinearController decelerator(static_cast<int>(speed), min_speed, accel_per_sec, updates_per_sec);
-
-    while ((degrees > 0 && GyroSingleton::instance()->get_accumulator() < degrees) ||
-           (degrees < 0 && GyroSingleton::instance()->get_accumulator() > degrees))
-    {
-        speed = decelerator.speed();
-        MOVEMENT_FUNCTION(static_cast<int>(speed * left_sign), static_cast<int>(speed * right_sign));
-        decelerator.step();
-        msleep(decelerator.get_msleep_time());
-    }
-
-    // stop at the end
-    MOVEMENT_FUNCTION(0, 0);
+    turn_degrees(&subscriber, max_speed, degrees, min_speed, accel_per_sec, updates_per_sec);
 }
